Organizes all layer components into seperate folder
