+++ app/board/[boardId]/_components/layers/layer-preview.tsx
